Fix date and time

+++ structures.h
@@ -34,8 +34,8 @@
 #include "types/tgl_message_media.h"
 
 std::shared_ptr<tgl_message> tglm_create_message(int64_t message_id, const tgl_peer_id_t& from_id,
-        const tgl_input_peer_t& to_id, const tgl_peer_id_t* fwd_from_id, const int* fwd_date,
-        const int* date, const std::string& message,
+        const tgl_input_peer_t& to_id, const tgl_peer_id_t* fwd_from_id, const int64_t* fwd_date,
+        const int64_t* date, const std::string& message,
         const tl_ds_message_media* media, const tl_ds_message_action* action,
         int32_t reply_id, const tl_ds_reply_markup* reply_markup, int flags);
 
@@ -43,7 +43,7 @@ std::shared_ptr<tgl_message> tglm_create_encr_message(const std::shared_ptr<tgl_
         int64_t message_id,
         const tgl_peer_id_t& from_id,
         const tgl_input_peer_t& to_id,
-        const int* date,
+        const int64_t* date,
         const std::string& message,
         const tl_ds_decrypted_message_media* media,
         const tl_ds_decrypted_message_action* action,
